Add support for /etc/wake-on-lan/special-wakeup.d
Server systems, like the big boxes, can't be woken up using
traditional wake-on-lan magic packets. For those systems, we plan to
use IPMI. The idea is to have an executable file, probably a simple
bash script, in /etc/wake-on-lan/special-wakeup.d/<hostname> that
triggers the IPMI wakeup for the particular host.